Move constants file and update imports
