+++ brats/core/docker.py
@@ -15,7 +15,7 @@ from rich.console import Console
 from rich.progress import Progress
 
 from brats.utils.algorithm_config import AlgorithmData
-from brats.utils.constants import DUMMY_PARAMETERS, PARAMETERS_DIR
+from brats.constants import DUMMY_PARAMETERS, PARAMETERS_DIR
 from brats.utils.exceptions import (
     AlgorithmNotCPUCompatibleException,
     BraTSContainerException,
