+++ brats/core/inpainting_algorithms.py
@@ -8,7 +8,7 @@ from typing import Optional
 from loguru import logger
 
 from brats.core.brats_algorithm import BraTSAlgorithm
-from brats.utils.constants import INPAINTING_ALGORITHMS, InpaintingAlgorithms, Task
+from brats.constants import INPAINTING_ALGORITHMS, InpaintingAlgorithms, Task
 from brats.utils.data_handling import input_sanity_check
